moving previous router patch to helpers. forked vue-router to introduce new patch which ignores double slash in url.
