+++ client/helpers/index.js
@@ -5,4 +5,5 @@ export { default as http } from './http';
 export { default as injectMomentDurationFormat } from './inject-moment-duration-format';
 export { default as jsonTryParse } from './json-try-parse';
 export { default as mapDomainDescription } from './map-domain-description';
+export { default as patchRouter } from './patch-router';
 export { default as shortName } from './short-name';
