+++ package.json
@@ -66,7 +66,7 @@
     "vue-js-modal": "^1.3.18",
     "vue-loader": "^13.0.1",
     "vue-prism-component": "^1.0.1",
-    "vue-router": "^3.0.1",
+    "vue-router": "github:just-at-uber/vue-router#61ee24d9a005451e7a5aedbd7a65582c68c3e15d",
     "vue-select": "nathanboktae/vue-select#after2.4",
     "vue-split-panel": "^1.0.4",
     "vue-style-loader": "^3.1.2",
